Address issue with potentially missing transaction. When reading the most recent transactions make sure we read them all: keep reading using the oldest of recently read signatures as value for "before" until we get fewer than sigFetchSize or zero (in this case try again in case provider returned zero in error) Add one more paraemter - maxReturn - to be used to limit number of signatures to keep. It is used in part B of low resource machine but can be used in general as well. These's inefficiency in this as we end up discarding some of the data we retrieve and we will have to retrieve it again, but this could be solved by storing data in a database.
For part C I change the way I choose data to store.
Initially I stored signatures that appeared at lest 2/3
of the time up to the first that didn't. Now I store all
signatures that appeat 2/3 of the time and discard others.
In production we can keep track of the discarded one and
retry them at a later time.